nextjs_host/model_server/requirements.txt
fastapi
uvicorn
python-chess
torch
tqdm

